Removed bigints

+++ sql/ad-hoc/2019-07-16 - Add sample todo data.sql
@@ -1,6 +1,6 @@
 DO $$
 DECLARE 
-    lastvalue BIGINT;
+    lastvalue INT;
 
 BEGIN
     TRUNCATE todolist CASCADE;
